Initial upload.
The new version of Unity broke my scripts, but they worked fine in one of the older versions. The classes are named in a fairly straightforward way, but of interest might be anything that has to do with Dialog. It's been a long time, but I remember setting it up so that I could have multiple dialog.asset files and link them to existing characters, while providing portrait and text data, and options for the player to select when it happens to be their turn in dialog.